Migrate all Triton kernels from make_block_ptr to make_tensor_descriptor

+++ flash_sparse_attn/ops/triton/quant.py
@@ -47,7 +47,7 @@ def quantize_fp8(x: torch.Tensor, dtype: torch.dtype = torch.float8_e5m2):
     FP8_MAX = 57344.0 if dtype == torch.float8_e5m2 else 448.0
 
     n_elements = x.numel()
-    amax = torch.zeros(1, device=x.device, dtype=x.dtype)
+    amax = torch.zeros(1, device=x.device, dtype=torch.float32)
     x_fp8 = torch.empty_like(x, dtype=dtype)
 
     BLOCK = 512
@@ -63,4 +63,4 @@ def quantize_fp8(x: torch.Tensor, dtype: torch.dtype = torch.float8_e5m2):
         x, x_fp8, scale, n_elements, BLOCK=BLOCK, FP8_DTYPE=tl_dtype
     )
 
-    return x_fp8, amax / FP8_MAX
+    return x_fp8, (amax / FP8_MAX).to(x.dtype)
